fix vote button

+++ app/helpers/application_helper.rb
@@ -1,6 +1,7 @@
 module ApplicationHelper
 
     def vote_unvote_btn(article, category)
+      if !current_user.nil?
         vote = Vote.find_by(article: article, user: current_user)
         if vote
           link_to('Unvote!', article_vote_path(id: vote.id, article_id: article.id, category_id: category), method: :delete, class:'cl-yl')
@@ -9,5 +10,6 @@ module ApplicationHelper
           link_to('Vote!', article_votes_path(article_id: article.id, user_id: current_user.id, category_id: category), method: :post, class:'cl-yl')
         end
     end
+  end
   
 end
